Backwards compatible deserialisation
This patch makes the deserialisation function backwards compatible
with the serialisation format prior to patch
https://github.com/dhil/mandriljulekalender/pull/1. As a result,
clients who has the calendar state stored in the previous format will
no longer crash.

This patch can be reverted after January 2024, as by this time every
client on the previous format will automatically be migrated to the
new format simply due to cookie expiration.